Move method for handling aggregation of distribution report docs in transaction service
