+++ server/src/worker/distribution-report-worker.ts
@@ -6,7 +6,6 @@ export function runDistributionReportingWorker(app: App, interval: string) {
     console.log(`Starting process of sending distribution reports at ${new Date()}...`);
     const result = await app.distributionReports.sendDistributionReportsToDonors();
     console.log(`Completed process of sending distribution reports at ${new Date()}`);
-    console.log(result);
     console.log();
   }, null, true, 'Africa/Nairobi');
